+++ mod.ts
@@ -1,2 +1,7 @@
 export { describe } from "./describe.ts";
-export { build } from "./build.ts";
+export {
+  build,
+  getRegisteredComponents,
+  isComponentRegistered,
+} from "./build.ts";
+export type { DescribeOptions, EventEmitter, Styles, Theme } from "./types.ts";